Testing how long updates take to reflect
As I'm at work (and got the ok to work on this), I've found that they don't allow access to github.  I'm going to try making updates within the github portal and I want to see how long it takes (from when I make and save a change...  to when the github-hosted version of my app) to update.  I'm commenting out the entirety of the css file to try this out.  When I have that number, I'm going to uncomment and notate how long it took.